Change searchPath
Using the default searchPath is ok when your connector doesn't have any implementation dependencies but when it does they are not included and your connector fails to load due to classes not found. By changing this to the install path, all the dependences are available and you are able to debug you connector with dependencies.